+++ Release Notes.md
@@ -31,6 +31,7 @@ SOFTWARE.
 * PWM: added case where duty cycle == scale in set_PWM_Duty_Scaled (to achieve real 100% duty cycles)
 * PWM: check if scale is equal 0 in set_PWM_Duty_Scaled (to avoid Division by 0)
 * GPIO_ex: added gpio logic and possible callbacks when input set / reset
+* GPIO_ex: removed eActOut (replaced by eGPIOstate in sarmfsw lib)
 * exceptions: removed erroneous dump_stack macro (functions will now be in stack_utils)
 * exceptions: dumpStack renamed to print_exception_stack, now static to exceptions.c
 * stack_utils: new module to get informations from global registers & stack